pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asset-manager"
version = "0.1.0"
description = "Texture library scanner, thumbnail generator, and Unity export tool"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=10.0",
    "Jinja2>=3.1",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.scripts]
assetmgr = "asset_manager.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
